Extracts new and update channel report modules from larger upsert module
